First Commit
Added poorly written and poorly commented C++ src file and some sample
pbm images to test with. Compile with eclipse cross C++ compiler and
add the image name in main C++ macro at start of the file.

Haven’t thought of anything much to with this project, but might clean
up the main.cpp in next commit.

Email me if you want to know what this program is doing.